add elasticsearch_state_storage api
