+++ .gitignore
@@ -4,7 +4,7 @@ logs
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
-
+.DS_Store
 # Runtime data
 pids
 *.pid
