+++ asset/elasticsearch_state_storage/interfaces.js
